requirements.txt
# Core AWS dependencies
boto3>=1.34.0
botocore>=1.34.0

# Development and testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
bandit>=1.7.0

# Utilities
pydantic>=2.0.0
click>=8.1.0
rich>=13.0.0
jinja2>=3.1.0

# Optional: Infrastructure as Code
aws-cdk-lib>=2.100.0
constructs>=10.0.0